+++ 1주차/HyungLee507/BOJ_1240.java
@@ -1,4 +1,4 @@
-package backjoon.study.week1;
+package backjoon.study.week1.HyungLee507;
 
 import java.io.BufferedReader;
 import java.io.IOException;
